requirements.txt
torch 
torchvision
opencv-python
flask 
insightface 
onnxruntime-gpu 
numpy
insightface
onnxruntime-gpu